donnai/requirements.txt
openai
plyer
pytesseract
Pillow
pyautogui
Node.js
NPM
Wine
electron
Django
djangorestframework
mysql-connector-python
requests
